+++ example/enhanced_change_notifier_example.dart
@@ -1,4 +1,5 @@
 import 'package:enhanced_change_notifier/enhanced_change_notifier.dart';
+import 'package:enhanced_change_notifier/signal.dart';
 
 class AppModel extends EnhancedChangeNotifier {
   Map<String, dynamic> _tasks = {};
@@ -26,6 +27,8 @@ class AppModel extends EnhancedChangeNotifier {
 
 void main() {
   final GlobalFactory<AppModel> appStateModel = GlobalFactory(() => AppModel());
+  Signal isConsumerReady = Signal();
+
   // add different types of listeners, support target, once, immediate
   appStateModel.getInstance().addListener(_e_appStateAnyChangedListener);
   appStateModel
@@ -40,6 +43,18 @@ void main() {
   appStateModel.getInstance().token = "fe3f6b58-684e-4063-ba3b-1b8f14981a8e";
   appStateModel.getInstance().baseUrl = "https://api.company.com";
   appStateModel.getInstance().tasks = {"task-for-example": 45};
+
+  isConsumerReady.value = false;
+
+  // delayed signal release
+  Future.delayed(Duration(milliseconds: 300), () {
+    isConsumerReady.value = true;
+  });
+
+  // register listener consumed immediately or awaited once via Signal(True).
+  isConsumerReady.promise(() => print("Task 1 executed"));
+  isConsumerReady.promise(() => print("Task 2 executed"));
+  isConsumerReady.promise(() => print("Task 3 executed"));
 }
 
 void _e_appStateAnyChangedListener() {