+++ lib/src/enhanced_value_notifier.dart
@@ -1,0 +1,49 @@
+import 'package:flutter/foundation.dart';
+import './enhanced_change_notifier.dart';
+
+/// A [EnhancedChangeNotifier] that holds a single value.
+///
+/// When [value] is replaced with something that is not equal to the old
+/// value as evaluated by the equality operator ==, this class notifies its
+/// listeners.
+///
+/// ## Limitations
+///
+/// Because this class only notifies listeners when the [value]'s _identity_
+/// changes, listeners will not be notified when mutable state within the
+/// value itself changes.
+///
+/// For example, a `EnhancedValueNotifier<List<int>>` will not notify its listeners
+/// when the _contents_ of the list are changed.
+///
+/// As a result, this class is best used with only immutable data types.
+///
+/// For mutable data types, consider extending [EnhancedChangeNotifier] directly.
+class EnhancedValueNotifier<T> extends EnhancedChangeNotifier
+    implements ValueListenable<T> {
+  /// Creates a [ChangeNotifier] that wraps this value.
+  EnhancedValueNotifier(this._value) {
+    if (kFlutterMemoryAllocationsEnabled) {
+      ChangeNotifier.maybeDispatchObjectCreation(this);
+    }
+  }
+
+  /// The current value stored in this notifier.
+  ///
+  /// When the value is replaced with something that is not equal to the old
+  /// value as evaluated by the equality operator ==, this class notifies its
+  /// listeners.
+  @override
+  T get value => _value;
+  T _value;
+  set value(T newValue) {
+    if (_value == newValue) {
+      return;
+    }
+    _value = newValue;
+    notifyListeners();
+  }
+
+  @override
+  String toString() => '${describeIdentity(this)}($value)';
+}