+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.7
+- Feature #1: Introduce `EnhancedValueNotifier` holds a single value.
+- Feature #2: Introduce `Signal` buffers pipelined listener pending a release signal.
+
 ## 0.1.6
 - README tuned
 
@@ -14,8 +18,8 @@
 - tiny code optimization 
 
 ## 0.1.1
-- Bug fixed #1: `target` listener supports VoidCallback and PropertyCallback
+- Bug fixed #1: `target` listener supports VoidCallback and PropertyCallback.
 
 ## 0.1.0
-- Feature #1: Support `target` `once` `immediate` for EnhancedChangeNotifier
-- Feature #2: GlobalFactory Class helps create a global singleton instance
+- Feature #1: Support `target` `once` `immediate` for EnhancedChangeNotifier.
+- Feature #2: GlobalFactory Class helps create a global singleton instance.